新增可转债仓位统计支持，新增仓位分类会把仓位归入A、HK、US等分类下

+++ utils.py
@@ -7,17 +7,28 @@ import time
 import json
 
 
+FOREIGN_ETF_CODELIST = ("513880", "513000", "159866", "513520", "164824", "159502", "159509", "501225")
+HK_ETF_CODELIST = ("513120", )
+
+ZH_CONVERTIBLE_BOND_CODE_SET_CACHE = None
+
 def isCodeAStock(code):
+	if not code.isdigit():
+		return False
 	code = str(code)
 	return len(code) == 6 and code[0] in ('0', '3', '6')
 
 
 def isCodeHKStock(code):
+	if not code.isdigit():
+		return False
 	code = str(code)
 	return len(code) == 5
 
 
 def isCodeETF(code):
+	if not code.isdigit():
+		return False
 	code = str(code)
 	return len(code) == 6 and code[0] in ('1', '5')
 
@@ -25,6 +36,14 @@ def isCodeETF(code):
 def isCodeCash(code):
 	code = str(code)
 	return code in ('CNY', 'HKD', 'USD')
+
+
+def isCodeZHConvertibleBond(code):
+	global ZH_CONVERTIBLE_BOND_CODE_SET_CACHE
+	if ZH_CONVERTIBLE_BOND_CODE_SET_CACHE is None:
+		codeDfRaw = getBasicData('zh_convertible_bond_code')
+		ZH_CONVERTIBLE_BOND_CODE_SET_CACHE = set(codeDfRaw['债券代码'])
+	return code in ZH_CONVERTIBLE_BOND_CODE_SET_CACHE
 
 
 def isCodeUSStock(code):
@@ -88,11 +107,31 @@ FOREIGN_EXCHANGE_DATA = None
 A_STOCK_DATA = None
 HK_STOCK_DATA = None
 ETF_DATA = None
+LOF_DATA = None
 US_STOCK_DATA = None
+ZH_CONVERTIBLE_BOND_DATA = None
+ZH_CONVERTIBLE_BOND_CODE = None
 OPEN_FUND_DAILY_DATA = None
 
 
+AKShareDataMap = {
+	"foreign_exchange_data": ('fx_spot_quote', 'data/foreign_exchange_data.pickle', FOREIGN_EXCHANGE_DATA),
+	"a_stock_data": ('stock_zh_a_spot_em', 'data/a_stock_data.pickle', A_STOCK_DATA),
+	"hk_stock_data": ('stock_hk_spot', 'data/hk_stock_data.pickle', HK_STOCK_DATA),
+	"etf_data": ('fund_etf_spot_em', 'data/etf_data.pickle', ETF_DATA),
+	"lof_data": ('fund_lof_spot_em', 'data/lof_data.pickle', LOF_DATA),
+	"us_stock_data": ('stock_us_spot', 'data/us_stock_data.pickle', US_STOCK_DATA),
+	"zh_convertible_bond_data": ('bond_zh_hs_cov_spot', 'data/zh_convertible_bond_data.pickle', ZH_CONVERTIBLE_BOND_DATA),
+	"zh_convertible_bond_code": ('bond_zh_cov_info_ths', 'data/zh_convertible_bond_code.pickle', ZH_CONVERTIBLE_BOND_CODE),
+	"open_fund_daily_data": ('fund_open_fund_daily_em', 'data/open_fund_daily_data.pickle', OPEN_FUND_DAILY_DATA),
+}
+
+clearedThisTime = False
 def clearCacheData():
+	global clearedThisTime
+	if clearedThisTime:
+		return
+	clearedThisTime = True
 	for path, dir_list, file_list in os.walk("data"):
 		for _file in file_list:
 			os.remove(os.path.join(path, _file))
@@ -101,21 +140,22 @@ def clearCacheData():
 		print("Cleared All file under data folder")
 
 
-def getBasicData(getFileKey=None):
-	global FOREIGN_EXCHANGE_DATA, A_STOCK_DATA, HK_STOCK_DATA, ETF_DATA, US_STOCK_DATA, OPEN_FUND_DAILY_DATA
+def cacheAllAKShareData():
+	for key, (funcName, filePath, cacheData) in AKShareDataMap.items():
+		print("Caching AKShare Data for %s" % key)
+		getBasicData(key, True)
 
-	aMap = {
-		"foreign_exchange_data": ('fx_spot_quote', 'data/foreign_exchange_data.pickle', FOREIGN_EXCHANGE_DATA),
-		"a_stock_data": ('stock_zh_a_spot_em', 'data/a_stock_data.pickle', A_STOCK_DATA),
-		"hk_stock_data": ('stock_hk_spot', 'data/hk_stock_data.pickle', HK_STOCK_DATA),
-		"etf_data": ('fund_em_etf_fund_daily', 'data/etf_data.pickle', ETF_DATA),
-		"us_stock_data": ('stock_us_spot', 'data/us_stock_data.pickle', US_STOCK_DATA),
-		"open_fund_daily_data": ('fund_open_fund_daily_em', 'data/fund_open_fund_daily_em.pickle', OPEN_FUND_DAILY_DATA),
-	}
+
+def getBasicData(getFileKey=None, force=False):
+	global FOREIGN_EXCHANGE_DATA, A_STOCK_DATA, HK_STOCK_DATA, ETF_DATA, LOF_DATA, US_STOCK_DATA, ZH_CONVERTIBLE_BOND_DATA, OPEN_FUND_DAILY_DATA
+	global AKShareDataMap
 
 	TIME_STAMP_FILE_PATH = "data/config.txt"
 
-	timeStr = time.strftime("%Y-%m-%d %H", time.localtime())
+	if force:
+		clearCacheData()
+
+	timeStr = time.strftime("%Y-%m-%d", time.localtime())
 	if os.path.exists(TIME_STAMP_FILE_PATH):
 		content = None
 		with open(TIME_STAMP_FILE_PATH) as cfgFile:
@@ -130,7 +170,7 @@ def getBasicData(getFileKey=None):
 			json.dump({"datetime": timeStr}, cfgFile, indent=4)
 
 	if getFileKey is not None:
-		funcName, localFileName, globalVal = aMap[getFileKey]
+		funcName, localFileName, globalVal = AKShareDataMap[getFileKey]
 		if globalVal is not None:
 			pass
 		elif not os.path.exists(localFileName):
